manifests: separate ingress tests in multiple test files to speed up ci
