+++ newsfragments/1144.internal.md
@@ -1,0 +1,1 @@
+CI: Separate the ingress manifests tests in 3 test runs.
